update to 1.4.4

+++ RawCraft/RawCraft/Network/Packets/ItemData.cs
@@ -9,7 +9,7 @@ namespace RawCraft.Network.Packets
         {
             s.ReadShort();
             s.ReadShort();
-            s.ReadData(s.ReadByte());
+            s.ReadData(s.ReadShort());
         }
     }
 }
